Clean up JiraClient configuration
Pulls JiraClient related code into the same folder. Paves the way for
accessing the version in JiraClient to allow for different API calls for
different versions when necessary.
