+++ src/jira/index.ts
@@ -1,2 +1,4 @@
 export * from './types';
 export * from './JiraClient';
+export * from './detectApiVersion';
+export { isValidJiraHost, buildJiraHostConfig } from './util';
